refactor(settings): Display new settings modal with regrouped modals as tabs.
