+++ src/lib/components/game/Atom.svelte
@@ -69,7 +69,7 @@
 </script>
 
 <button
-	class="atom relative mt-20 flex size-112.5 items-center justify-center cursor-pointer bg-transparent md:size-90 sm:size-75"
+	class="atom relative mt-20 flex size-64 sm:size-75 md:size-90 lg:size-112.5 items-center justify-center cursor-pointer bg-transparent"
 	class:bonus={gameManager.hasBonus}
 	onclick={async e => await handleClick(e)}
 	bind:this={atomElement}
