+++ src/lib/components/layout/NavBar.svelte
@@ -1,9 +1,5 @@
 <script lang="ts">
-	import Changelog from '@components/modals/Changelog.svelte';
-	import CloudSave from '@components/modals/CloudSave.svelte';
-	import Credits from '@components/modals/Credits.svelte';
-	import FeedbackForm from '@components/modals/FeedbackForm.svelte';
-	import GlobalStats from '@components/modals/GlobalStats.svelte';
+	import Settings from '@components/modals/Settings.svelte';
 	import Leaderboard from '@components/modals/Leaderboard.svelte';
 	import SkillTree from '@components/modals/SkillTree.svelte';
 	import Electronize from '@components/prestige/Electronize.svelte';
@@ -13,9 +9,11 @@
 	import { changelog } from '$stores/changelog';
 	import { gameManager } from '$helpers/GameManager.svelte';
 	import { remoteMessage } from '$stores/remoteMessage.svelte';
+	import { ui } from '$stores/ui.svelte';
 	import { mobile } from '$stores/window.svelte';
-	import { ChartNoAxesColumn, Network, Info, Atom, Trophy, MessageSquare, Orbit, FileText, Cloud, type Icon as IconType } from 'lucide-svelte';
+	import { Network, Atom, Trophy, Orbit, Settings as SettingsIcon, type Icon as IconType } from 'lucide-svelte';
 	import { onDestroy, onMount, type Component } from 'svelte';
+
 
 	type NavBarComponent = Component<{ onClose: () => void }>;
 
@@ -28,11 +26,6 @@
 	}
 
 	const links: Link[] = [
-		{
-			icon: ChartNoAxesColumn,
-			label: 'Stats',
-			component: GlobalStats,
-		},
 		{
 			icon: Trophy,
 			label: 'Leaderboard',
@@ -60,38 +53,28 @@
 			notification: () => gameManager.electronizeElectronsGain > 0,
 		},
 		{
-			icon: FileText,
-			label: 'Changelog',
-			component: Changelog,
+			icon: SettingsIcon,
+			label: 'Parameters',
+			component: Settings,
 			notification: () => $changelog.hasUnread,
-		},
-		{
-			icon: Cloud,
-			label: 'Cloud Save',
-			component: CloudSave,
-		},
-		{
-			icon: Info,
-			label: 'Credits',
-			component: Credits,
-		},
-		{
-			icon: MessageSquare,
-			label: 'Feedback',
-			component: FeedbackForm
 		},
 	];
 
-	let activeComponent: NavBarComponent | null = $state(null);
+	/* Filter out Settings from main links - we want it always at the bottom */
+	const mainLinks = links.filter(l => l.component !== Settings);
+	const settingsLink = links.find(l => l.component === Settings)!;
+
 	let visibleComponents: Link[] = $state([]);
 
 	let interval: ReturnType<typeof setInterval> | null = null;
 
 	onMount(() => {
-		visibleComponents = links.filter(link => !link.condition || link.condition());
-		interval = setInterval(() => {
-			visibleComponents = links.filter(link => !link.condition || link.condition());
-		}, 100);
+		ui.registerSettings(Settings);
+		const updateVisible = () => {
+			visibleComponents = mainLinks.filter(link => !link.condition || link.condition());
+		};
+		updateVisible();
+		interval = setInterval(updateVisible, 100);
 	});
 
 	onDestroy(() => {
@@ -101,22 +84,36 @@
 
 {#if mobile.current}
 	<div
-		class="absolute top-[33vh] -translate-y-1/2 z-10 grid gap-3.5 w-full justify-between pointer-events-none"
-		class:grid-cols-2={visibleComponents.length > 5}
-		class:px-2={visibleComponents.length > 5}
-		class:left-4={visibleComponents.length < 5}
-		style:grid-template-columns={visibleComponents.length > 5 ? 'auto auto' : 'auto'}
+		class="absolute top-[33vh] -translate-y-1/2 z-10 grid gap-3.5 pointer-events-none"
+		class:grid-cols-2={visibleComponents.length + 1 >= 5}
+		class:inset-x-0={visibleComponents.length + 1 >= 5}
+		class:justify-between={visibleComponents.length + 1 >= 5}
+		class:left-4={visibleComponents.length + 1 < 5}
+		class:px-4={visibleComponents.length + 1 >= 5}
+		class:w-full={visibleComponents.length + 1 >= 5}
+		style:grid-template-columns={visibleComponents.length + 1 >= 5 ? 'auto auto' : 'auto'}
 	>
 		{#each visibleComponents as link}
 			<NotificationDot hasNotification={link.notification ? link.notification() : false}>
 				<button
 					class="flex items-center justify-center rounded-lg bg-accent/90 p-2 text-white transition-all hover:bg-accent pointer-events-auto"
-					onclick={() => activeComponent = link.component}
+					onclick={() => ui.openModal(link.component)}
 				>
 					<link.icon size={30} />
 				</button>
 			</NotificationDot>
 		{/each}
+
+		<!-- Mobile Settings (add to grid or place separately? User said bottom of navbar, which implies desktop mostly, but let's add it here too if space permits or just keep it in flow) -->
+		<!-- For mobile, just append it to the list effectively -->
+		<NotificationDot hasNotification={settingsLink.notification ? settingsLink.notification() : false}>
+			<button
+				class="flex items-center justify-center rounded-lg bg-accent/90 p-2 text-white transition-all hover:bg-accent pointer-events-auto"
+				onclick={() => ui.openModal(settingsLink.component)}
+			>
+				<settingsLink.icon size={30} />
+			</button>
+		</NotificationDot>
 	</div>
 {:else}
 	<nav
@@ -127,23 +124,41 @@
 			<NotificationDot hasNotification={link.notification ? link.notification() : false}>
 				<button
 					class="group relative flex h-12 w-12 items-center justify-center rounded-lg bg-accent/90 text-white transition-all hover:bg-accent"
-					onclick={() => activeComponent = link.component}
+					onclick={() => ui.openModal(link.component)}
 				>
 					<link.icon size={32} />
 					<span
-						class="label invisible absolute left-[calc(100%+1.25rem)] whitespace-nowrap rounded-lg bg-accent/90 px-3 py-2 text-sm opacity-0 transition-all group-hover:visible group-hover:opacity-100 bg-accent-900"
+						class="label invisible absolute left-[calc(100%+1.25rem)] whitespace-nowrap rounded-lg bg-accent/90 px-3 py-2 text-sm opacity-0 transition-all group-hover:visible group-hover:opacity-100 bg-accent-900 border border-white/10 shadow-xl z-50"
 					>
 						{link.label}
 					</span>
 				</button>
 			</NotificationDot>
 		{/each}
+
+		<div class="flex-1"></div>
+
+		<NotificationDot hasNotification={settingsLink.notification ? settingsLink.notification() : false}>
+			<button
+				class="group relative flex h-12 w-12 items-center justify-center rounded-lg bg-accent/90 text-white transition-all hover:bg-accent"
+				onclick={() => ui.openModal(settingsLink.component)}
+			>
+				<div class="transition-transform duration-500 group-hover:rotate-90">
+					<settingsLink.icon size={32} />
+				</div>
+				<span
+					class="label invisible absolute left-[calc(100%+1.25rem)] whitespace-nowrap rounded-lg bg-accent/90 px-3 py-2 text-sm opacity-0 transition-all group-hover:visible group-hover:opacity-100 bg-accent-900 border border-white/10 shadow-xl z-50"
+				>
+					{settingsLink.label}
+				</span>
+			</button>
+		</NotificationDot>
 	</nav>
 {/if}
 
-{#if activeComponent}
-	{@const SvelteComponent = activeComponent}
-	<SvelteComponent onClose={() => (activeComponent = null)} />
+{#if ui.activeModal}
+	{@const SvelteComponent = ui.activeModal}
+	<SvelteComponent onClose={() => ui.closeModal()} />
 {/if}
 
 <style>
